Moved tab class, updated close method
