+++ requirements/tests.txt
@@ -1,2 +1,2 @@
-pytest==3.0.6
+pytest==3.0.7
 pytest-selenium==1.9.1
